角色的增删改查

+++ server/model/model.go
@@ -61,5 +61,5 @@ type ResPageData struct {
 
 // 根据id查询接参结构
 type GetById struct {
-	ID int `json:"id" uri:"id" form:"id" binding:"required"` // 主键ID
+	ID uint64 `json:"id" uri:"id" form:"id" binding:"required"` // 主键ID
 }
